requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.2
Pillow>=10.1.0
bcrypt>=4.1.2
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.1.4